AI maneuvers take stress into account.
The AI now uses the TokenModule to count its stress and picks from the stress table if appropriate. It sheds stress on a blue maneuver and adds stress on a red manveuver. Actions are not implemented yet, but the situations that prevent actions from being taken are being recorded and considered.

I do think that ultimately the stress will have to be player-specified, as the tokens can get mixed up when ships pass close to each other. This isn't a problem for player ships, as it can be discussed and corrected, but the AI ships don't have that luxury. We'll have to playtest whole games and decide.