Start of AI implementation (#118)
* Split the display part of CheckArc into a separate function.

The logic has been moved into a separate function, ArcCheck.GetTargetsInRelationToArc. This will allow it to be reused by the AI in the future.

* Sort targets returned from CheckArc by desirability

In-arc targets are displayed first, nearest to most distant, followed by out of arc targets. Targets at the same range are ordered by obstruction.

* Initial commit of the AI module

A rough partial implementation of the core of the AI system. Functionality so far consists of AI ships selecting a target, planning an executing a move, and avoiding obstacles. Only includes data for the TIE/ln Fighter.
To do list includes: stress accumulation, target lock handling, 0-range enemies, table edge avoidance, action selection, squads, different AI types (strike, flee, protect, etc), and moving the interface from the context menu to a dial.

* AI maneuvers take stress into account.

The AI now uses the TokenModule to count its stress and picks from the stress table if appropriate. It sheds stress on a blue maneuver and adds stress on a red manveuver. Actions are not implemented yet, but the situations that prevent actions from being taken are being recorded and considered.

I do think that ultimately the stress will have to be player-specified, as the tokens can get mixed up when ships pass close to each other. This isn't a problem for player ships, as it can be discussed and corrected, but the AI ships don't have that luxury. We'll have to playtest whole games and decide.

* Added all Imperial AIs. Some target selections are still TODO

Also added a function to check the validity of AI move tables.

* Disable the AI until the dial UI is implemented.