AI takes ion tokens into account
The AI responds to its current ion count, and will select an s1 maneuver if it has the relevant number of tokens. It also removes them afterwards, and treats the maneuver as a blue.

There is a bug that I discovered while testing - when removing the second token from a token stack, the stack becomes a 'generic' type for one frame before reverting again to a 'token' type. This means that the ship doesn't pick up the stack and have it follow it along to its new location. Easiest way to repro this is put two stress tokens on a ship and perform a blue maneuver.