Add new action system, actions for four ships.
Major rework of the action system! It is now modular rather than based on bespoke functions, so suppports flexible definitions and condition checking.

This commit also includes the actions for four Imperial ships: the Fighter, Interceptor, Advanced x1, and the Bomber.

Note that the target lock actions aren't fully implemented, as they aren't automatically assigned to the target ship. This will be coming soon.