Add actions AI for TIE Phantom
Added the AI action behaviour for the TIE Phantom, included support for cloak actions.

There is an outstanding bus: cloak tokens added through AI or through the dial do attach to ships okay, but they don't spawn the Delete and Decloak buttons. TokenModule.QueueShipTokensMove picks up on this and excludes them from moving with the ship.