Add ability to query and set the arcs on indicators.
This will allow the AI to check its turret locations and change them when necessary. Also it allows the arc check to automatically display the correct arc instead of having all the possible options in the context menu.
The link between turrets and ships is stored across save games.

Outstanding issues:
- Huge ships still only have one turret mount, in the middle of the ship. Code can be added in Arc Indicator.ttslua to build in support for this, I've left appropriate TODO comments in these areas.
- While the link between indicators and ships is kept across save games, the link between tokens is not. This means that arc indicator tokens, just like all other tokens, won't follow a ship after a saved game has been loaded. This should wait for a subsequent commit that will deal with all tokens.