Support for multiple mount points for arc indicators
This commit defines the number and positions of turret mount points for differently-sized ships, and allows multiple arc indicators per ship.

Note that this doesn't spawn the correct number of arc indicators for huge ships with more than one turret upgrade, this is separate task outside of this commit's scope.