AI support for turrets, Decimator actions
This commit allows the AI to control its turrets by checking for targets in different arcs and rotating to ones that have valid targets.

It also now waits for all of the AI ship's tokens to come to a rest before taking an action, otherwise the arc indicator won't have rotated and it is out of sync with the ship.